+++ Assets/_Build/Scripts/Game/GameData/Settings/SettingBars/SliderSettingDataSO.cs
@@ -5,7 +5,7 @@ namespace LostKaiju.Game.GameData.Settings
     [CreateAssetMenu(fileName = "SliderSettingDataSO", menuName = "Scriptable Objects/Settings/SliderSettingDataSO")]
     public class SliderSettingDataSO : SettingBarDataSO, ISliderSettingData
     {
-        [field: SerializeField] public float MinValue { get; private set; }
-        [field: SerializeField] public float MaxValue { get; private set; }
+        [field: SerializeField] public int MinValue { get; private set; }
+        [field: SerializeField] public int MaxValue { get; private set; }
     }
 }
